changed leds for the nrf stuff, moved the task to our own code directory
