+++ BaseStation_stm32f4discovery/Src/robocup/hermes/communicationTask.h
@@ -1,0 +1,13 @@
+/*
+ * communicationTask.h
+ *
+ *  Created on: Apr 19, 2017
+ *      Author: znuxor
+ */
+
+#ifndef ROBOCUP_HERMES_COMMUNICATIONTASK_H_
+#define ROBOCUP_HERMES_COMMUNICATIONTASK_H_
+
+
+
+#endif /* ROBOCUP_HERMES_COMMUNICATIONTASK_H_ */
